Implement audio endpoint
Added method for posting audio data tied to a particular query,
according to documentation from Flask for uploading files. Has not
yet been thoroughly tested, but cursory tests show that it accepts
a file from curl, stores it, and correctly logs the path in the
database for the entry of the corresponding query.