Accept wavs as audio, and require 3 or more tokens as CNN input
Taking wavs as input to the audio endpoint simplifies things on
the client side, especially as compared with the relatively small
benefit of compression compared to computational cost.

The other change corrects a bug in which queries were lost due to
the CNN throwing errors if receiving input that didn't span at
least the width of one kernel.