General scoring support
Added score endpoint to webservice.

Added pdf rendering support to static files, via jsPDF.

Added template for rendering the pdf to a browser.

Corrected the database query in GET /conversations/ to correctly
pass an iterable item as the second parameter to cursor.execute().

Now storing the full transcript summary string returned by Chat-
Script in the database when "score me" is sent as the content of
a query.

Added Raw_score field to Conversations table in the database
specification; this is stored as a TEXT, allowing 64kb, but without
indexing.

Corrected a 404 Error response that was crashing the server.